pyproject.toml
[project]
name = "graphrag-neo4j"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fpdf>=1.7.2",
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-experimental>=0.3.4",
    "langchain-neo4j>=0.4.0",
    "langchain-ollama>=0.3.0",
    "langchain-openai>=0.3.11",
    "neo4j>=5.28.1",
    "neo4j-graphrag>=1.6.1",
    "openai>=1.68.2",
    "pydantic>=2.10.6",
    "pydantic-core>=2.27.2",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.0",
]
